+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ next
 - Require registration in ``mpl.rcParams["figure.hooks"]`` for the
   :envvar:`MPLCURSORS` environment variable.
 - Support new-style (Matplotlib≥3.8) contour plots.
+- Support :class:`~matplotlib.image.BboxImage`.
 
 0.6
 ---